Add a scanable QR code to the bitcoin return
This is completely untested, and you've been warned. I wrote this in the github inline editor. 

However, in theory, it should spit out the QR code in an image tag alongside the bitcoin address
As a side note, you'll need to install io and qrcode, and possibly install Pillow instead of PIL